Internal debug aid only: log to RAM.
To aid debugging of detailed real-time issues a RAM logging facility is added.  This is not intended for normal use, is not included in ubxlib.h, does not form a part of the normal ubxlib code. etc.; it is only used when required during development to debug those awkward, usually platform-related, real-time issues, the ones where break-pointing won't help.